+++ src/components/App/App.css
@@ -1,0 +1,15 @@
+* {
+  box-sizing: border-box;
+}
+
+html {
+  font-size: 10pt;
+  line-height: 1.4;
+  font-weight: 400;
+  font-family: "Source Sans Pro", "Open Sans", Roboto, Arial, sans-serif;
+}
+
+body {
+  margin: 0;
+  background: #eee;
+}
