+++ src/components/App/App.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Welcome from '../Welcome/Welcome';
+import Header from '../Header/Header';
+import Article from '../Article/Article';
+
+import './App.css';
+
+function App() {
+  return (
+    <main className="app">
+      <Welcome />
+      <Header />
+      <Article />
+    </main>
+  );
+}
+
+export default App;
