+++ src/components/Header/Header.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Navigation from '../Navigation/Navigation';
+
+import './Header.css';
+
+function Header() {
+  return (
+    <header className="header">
+      <h1 className="header__title">Site Name</h1>
+      <Navigation />
+    </header>
+  );
+}
+
+export default Header;
