+++ src/components/Welcome/Welcome.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './Welcome.css';
+
+function Welcome() {
+  return (
+    <section className="welcome">
+      <span className="welcome__text">Sticky Header!</span>
+    </section>
+  );
+}
+
+export default Welcome;
